+++ assets/i18n/de.json
@@ -112,8 +112,10 @@
   "amsco_key_hint": "Fortlaufenden Ziffern von 1 bis 9",
   "amsco_error_key": "Schlüssel fehlerhaft!",
   "amsco_usedgrid": "Verwendete Matrix",
-  "animated_gif_title": "Animated Gif's",
-  "animated_gif_description": "Animierte Gif's analysieren",
+  "animated_image_title": "Animierte Bilder",
+  "animated_image_description": "Animierte Bilder analysieren (gif, png, webp)",
+  "animated_image_table_index": "Index",
+  "animated_image_table_duration": "Dauer in ms",
   "apparenttemperature_selection_title": "Gefühlte Temperatur",
   "apparenttemperature_selection_description": "Berechnet den Wohlfühlfaktor in Abhängigkeit von Temperatur, Luftfeuchtigkeit und Wind",
   "asciivalues_title": "ASCII und Unicode (UTF-8)",
