+++ lib/widgets/searchstrings/searchstrings_common.dart
@@ -4,7 +4,7 @@ const SEARCHSTRING_COMMON_ADFGVX = 'adfgx adfgvx polybius polybios transposition
 const SEARCHSTRING_COMMON_AFFINE = 'affine';
 const SEARCHSTRING_COMMON_ALPHABETVALUES = '';
 const SEARCHSTRING_COMMON_AMSCO = 'amsco transposition spaltentausch swap columns';
-const SEARCHSTRING_COMMON_ANIMATED_GIFS = 'gif jpg';
+const SEARCHSTRING_COMMON_ANIMATED_IMAGES = 'gif png webp';
 const SEARCHSTRING_COMMON_APPARENTTEMPERATURE = '';
 const SEARCHSTRING_COMMON_APPARENTTEMPERATURE_HEATINDEX = '';
 const SEARCHSTRING_COMMON_APPARENTTEMPERATURE_HUMIDEX = 'humidex ';
