+++ lib/widgets/utils/file_utils.dart
@@ -22,6 +22,7 @@ Map<List<int>, String> _fileTypes = {
   [0x42, 0x4D] : ".bmp",
   [0x47, 0x49, 0x46, 0x38, 0x39, 0x61] : ".gif",
   [0x47, 0x49, 0x46, 0x38, 0x37, 0x61] : ".gif",
+  [0x52, 0x49, 0x46, 0x46] : ".webp",
   [0x30, 0x26, 0xB2, 0x75] : ".wmv",
   [0x49, 0x44, 0x33, 0x2E] : ".mp3",
   [0x49, 0x44, 0x33, 0x03] : ".mp3",
